[TASK] Remove compat layer for new 6.0 classes - 2
Remove the name space compatibility layer for new 6.0 classes in
t3lib/. This second patch finishes the work started with #40240
and removes the compatibily layer for the log API, and the new
configuration and rootline handling.
All class occurences that still referenced the old class names
in method headers and comments are adapted.

Change-Id: Icb49db084dc25ed7b1a0b2e092c3b7a4e672b01c
Resolves: #40244
Related: #40095
Related: #40240
Releases: 6.0
Reviewed-on: http://review.typo3.org/14122
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn