[TASK] Namespace and move t3lib unit tests - 1
This patch moves most unit tests of the tests/Unit/t3lib/ main
directory. They follow the new class names of the classes under
test and are namespaced. Covered are files that belong to the
backend and core extension.
The travis test configuration is adapted to execute tests
located in those extension folders.

Change-Id: Ibda0cb72596a418a3f97eeba83f39428a4d0d532
Resolves: #40247
Related: #40095
Releases: 6.0
Reviewed-on: http://review.typo3.org/14123
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn