[BUGFIX] Install process fails on case-insensitive file systems
The install process fails on case-insensitive file systems like
HFS+ on Mac OS X. The reason is that "classes" and "Classes" are
the same in that regard, which will load the old saltedpasswords
autoloader. Since the extension is not enabled in the install
process, this will fail of course.

Change-Id: Ic4e9a7d4149faeb5b68ae238101d27a10c5d5f7e
Fixes: #40438
Releases: 6.0
Reviewed-on: http://review.typo3.org/14228
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader