[TASK] Clean up extensionmanager test directory
There are still those empty 'old' test files from the namespace script
that moved the tests to the Tests/Unit subfolder. They are removed now.
Additionally the test class comments to state the name of the class
under test are removed, this information is redundant and is obvious
from the test class and file name.

Change-Id: I0ef68252543e058cb76137518276fa2db4f9febd
Resolves: #40517
Related: #40095
Releases: 6.0
Reviewed-on: http://review.typo3.org/14277
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn