[BUGFIX] Usage of raw binary characters in SQL Parser
Currently, the SQL Parser uses raw binary characters. In some cases,
this breaks the PHP interpreter. This results in the file
SqlParser.php not being parsed (the PHP interpreter just dumps
"???..." to stdout) and a follow-up fatal error because the class
cannot be found.

This patch replaces the raw binary characters with corresponding
character sequence in double quotes.

Change-Id: I28fb5d0c8401794e8f4bf20e7e2611b4cf7014ae
Fixes: #40932
Relates: #40672
Releases: 6.0
Reviewed-on: http://review.typo3.org/14689
Reviewed-by: Oliver Hader
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel